# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "记账 API。部分业务错误使用 HTTP 200，通过 status/msg 判断结果。",
        "title": "Jiezhang API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/super_chart/categories_list": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "分类 Top 排行，Query: `year`, `month`",
                "operationId": "SuperChartHandler_CategoriesList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.SuperCategoryTopItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/super_chart/get_pie_data": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "饼图数据，Query: `year`, `month`, `statement_type`",
                "operationId": "SuperChartHandler_PieData",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "statement_type",
                        "name": "statement_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.SuperPieItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/super_chart/header": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "图表头部数据，Query: `year`, `month`",
                "operationId": "SuperChartHandler_Header",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.SuperChartHeaderData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/super_chart/line_chart": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "折线图数据，Query: `year`, `month`",
                "operationId": "SuperChartHandler_LineChart",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.SuperLineChartData"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/super_chart/table_sumary": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "日汇总表，Query: `year`, `month`",
                "operationId": "SuperChartHandler_TableSummary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.SuperTableSummaryItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
        "/super_chart/week_data": {
            "get": {
                "security": [
                    {
                        "AppID": [],
                        "SessionKey": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SuperChart"
                ],
                "summary": "周数据，Query: `year`, `month`",
                "operationId": "SuperChartHandler_WeekData",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本 ID；未提供时使用用户默认账本",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "month",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功；兼容接口也可能在 HTTP 200 的 status/msg 中返回业务错误",
                        "schema": {
                            "$ref": "#/definitions/types.SuperWeekData"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    },
                    "500": {
                        "description": "服务错误",
                        "schema": {
                            "$ref": "#/definitions/types.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "session": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "types.SuperCategoryTopItem": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "data": {
                    "type": "number"
                },
                "format_amount": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "percent": {
                    "type": "string"
                }
            }
        },
        "types.SuperChartHeaderData": {
            "type": "object",
            "properties": {
                "expend_count": {
                    "type": "number"
                },
                "expend_percent": {
                    "type": "number"
                },
                "expend_rise": {
                    "type": "string"
                },
                "income_count": {
                    "type": "number"
                },
                "income_percent": {
                    "type": "number"
                },
                "income_rise": {
                    "type": "string"
                },
                "surplus": {
                    "type": "number"
                },
                "surplus_percent": {
                    "type": "number"
                },
                "surplus_rise": {
                    "type": "string"
                }
            }
        },
...
        "types.SuperLineChartData": {
            "type": "object",
            "properties": {
                "expends": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "incomes": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "months": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "surplus": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "types.SuperPieItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "types.SuperTableSummaryItem": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "expend": {
                    "type": "string"
                },
                "income": {
                    "type": "string"
                },
                "total_expend": {
                    "type": "string"
                },
                "total_income": {
                    "type": "string"
                },
                "total_surplus": {
                    "type": "string"
                }
            }
        },
...
        "types.SuperWeekData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "weeks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppID": {
            "type": "apiKey",
            "name": "X-WX-APP-ID",
            "in": "header"
        },
        "SessionKey": {
            "type": "apiKey",
            "name": "X-WX-Skey",
            "in": "header"
        }
    }
}